split FavoritePhraseLoader for bundle lazyload
