+++ demos/pet-demo/src/components/__isograph/Pet/FavoritePhraseLoader/output_type.ts
@@ -1,4 +1,4 @@
 import type { ExtractSecondParam, CombineWithIntrinsicAttributes } from '@isograph/react';
 import type React from 'react';
-import { FavoritePhraseLoader as resolver } from '../../../FavoritePhrase';
+import { FavoritePhraseLoader as resolver } from '../../../FavoritePhraseLoader';
 export type Pet__FavoritePhraseLoader__output_type = (React.FC<CombineWithIntrinsicAttributes<ExtractSecondParam<typeof resolver>>>);
