+++ demos/pet-demo/src/components/__isograph/Pet/FavoritePhraseLoader/resolver_reader.ts
@@ -1,6 +1,6 @@
 import type {ComponentReaderArtifact, ExtractSecondParam, ReaderAst } from '@isograph/react';
 import { Pet__FavoritePhraseLoader__param } from './param_type';
-import { FavoritePhraseLoader as resolver } from '../../../FavoritePhrase';
+import { FavoritePhraseLoader as resolver } from '../../../FavoritePhraseLoader';
 
 const readerAst: ReaderAst<Pet__FavoritePhraseLoader__param> = [
   {
